refactor: package 구조 수정 - webser package 내부로 utils 패키지 이동 - request, response package를 webserver에 위치할 수 있도록 수정 - statuscode, contenttype Enum들을 각각의 패키지로 구성
