+++ src/main/java/webserver/response/StaticFileResponse.java
@@ -1,4 +1,4 @@
-package webserver.handlers.response;
+package webserver.response;
 
 public class StaticFileResponse extends HttpResponse {
