+++ src/test/java/utils/RequestUtilsTest.java
@@ -2,6 +2,7 @@ package utils;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import webserver.utils.RequestUtils;
 
 import static org.assertj.core.api.Assertions.*;
